Add files via upload
This file contains a list of CTL epitopes that were further processed to determine those that are most suitable for vaccine construction. The analysis of these epitopes was based on the 10 different alleles used during MHC I prediction. From the 10 alleles used HLA-B*07:02 was the only one that did not have an epitopes that could bind to it